Update to compile on Zig 0.13
It still does not link because version 0.6.0 of wgpu-native that we are
vendoring does not have a prebuilt binary for macOS ARM64, only x86_64.
Two options would be to figure out how to link an x86_64 library to an
ARM64 target in build.zig, or update to a newer version of wgpu-native
that does ship the ARM64 variant of the library. I already looked into
the latter but later versions have non-trivial API changes, so doing
that will be more work still.